cedula is now bigint

+++ src/database/query.sql
@@ -19,7 +19,7 @@ USE `ferreteria-construmanta-p`;
 DROP TABLE IF EXISTS `ferreteria-construmanta-p`.`usuarios`;
 CREATE TABLE IF NOT EXISTS `ferreteria-construmanta-p`.`usuarios` (
   `id` CHAR(36) NOT NULL DEFAULT (UUID()),
-  `cedula` INT NOT NULL,
+  `cedula` BIGINT NOT NULL,
   `nombre_completo` VARCHAR(255) NOT NULL,
   `correo_electronico` VARCHAR(255) NOT NULL,
   `telefono` VARCHAR(24) NOT NULL,
